Refactoring Chat system to break out components and optimze things
The current goal is to hybridize the group and chat pages, breaking
out these components so that we can make that happen in a somewhat clean
way. We're also doing some chat page optimizations on the way to fix
some performance issues and in general clean things up and progress
things towards the new react hook stuff.
