+++ src/components/Chat/Chat.styl
@@ -22,9 +22,6 @@ channel-width=12rem;
     position: relative;
     display: flex;
     min-height: 20rem;
-    border-left: 1px solid transparent;
-    border-top: 1px solid transparent;
-    themed border-color shade3
     .input-container {
         border-bottom: 1px solid transparent;
     }
@@ -54,106 +51,9 @@ channel-width=12rem;
     .all-channels {
         display: flex;
         flex-direction: column;
-        height: "calc(100vh - %s - %s)" % (navbar-height channel-utils-height);
+        height: 100vh;
         overflow-y: auto;
         width: channel-width;
-    }
-    .channels {
-        flex-grow: 0;
-        flex-shrink: 0;
-        overflow-x: hidden;
-
-        .channel-header {
-            margin-top: 1.0rem;
-            text-transform: uppercase;
-            font-weight: bold;
-            font-size: font-size-smaller;
-            margin-bottom: 0.3rem;
-            display: flex;
-            justify-content: space-between;
-            align-content: center;
-
-            .channel-expand-toggle {
-                padding: 0.2rem;
-                padding-right: 0.5rem;
-                cursor: pointer;
-
-            }
-        }
-        .channel-header:first-child {
-            margin-top: 0;
-        }
-
-        .channel {
-            padding-left: 0.2rem;
-            display: flex;
-            justify-content: space-between;
-            align-items: center;
-
-            &.active {
-                themed background-color primary
-                themed color colored-background-fg
-            }
-            &.unjoined {
-                themed color shade2
-            }
-            &.unread {
-                themed color primary
-            }
-            &.mentioned {
-                themed color chat-mentions
-            }
-        }
-
-        .channel-name {
-            flex-shrink: 1;
-            flex-grow: 1;
-            cursor: pointer;
-            overflow: hidden;
-            white-space: nowrap;
-        }
-
-        .unread-count {
-            flex-shrink: 0;
-            @extend .monospace
-            font-size: 0.8rem;
-            font-style: normal !important;
-            float: right;
-            padding-left: 0.5rem;
-            padding-top: 0.10rem;
-
-            &:before {
-                content: attr(data-count);
-            }
-        }
-        .channel:hover .unread-count:before {
-            cursor: pointer;
-            content: attr(data-menu);
-            border: 1px solid transparent
-            border-radius: 0.2rem;
-            themed background-color shade4
-            themed border-color shade3
-            padding-left: 0.3rem;
-            padding-right: 0.3rem;
-            themed color fg
-        }
-        .channel:hover .unread-count:hover:before {
-            themed background-color shade5
-            themed border-color shade4
-            themed color fg
-        }
-    }
-    .channels.hide-unjoined {
-        .unjoined {
-            display: none;
-        }
-    }
-    .seekgraph-container {
-        height: channel-utils-height;
-        .SeekGraph {
-            width: channel-width;
-            height: 8rem;
-        }
     }
     .channel-buttons {
         display: flex;
@@ -181,17 +81,7 @@ channel-width=12rem;
         }
 
     }
-    .Player {
-        themed color fg
-    }
     .center-container {
-        flex-basis: 100%;
-        flex-grow: 1;
-        flex-shrink: 1;
-        display: flex;
-        flex-direction: column;
-        border-right: 1px solid transparent;
-        themed border-right-color shade3
 
         .chat-log {
             flex-basis: 100%;
@@ -214,7 +104,39 @@ channel-width=12rem;
             padding: 0.2rem;
         }
     }
+}
 
+
+
+
+
+
+.EmbeddedChat, .ChatUsersList {
+    .Player {
+        themed color fg
+        font-weight: bold;
+    }
+}
+
+.EmbeddedChat {
+    display: flex;
+    flex-basis: 100%;
+    flex-grow: 1;
+    flex-shrink: 1;
+    flex-direction: column;
+
+
+    .chat-log {
+        flex-basis: 100%;
+        flex-grow: 1;
+        flex-shrink: 1;
+        overflow-y: scroll;
+        border-bottom: 1px solid transparent;
+        themed border-bottom-color shade3
+        .Player {
+            themed color fg
+        }
+    }
 
     .timestamp {
         @extend .monospace
@@ -237,10 +159,6 @@ channel-width=12rem;
         themed color chat-self-message
     }
 
-    .icon {
-        width: 15px;
-        height: 15px;
-    }
 
     .system {
         themed color chat-system
@@ -251,11 +169,6 @@ channel-width=12rem;
         padding-bottom: 0.5rem
         margin-top: 0.5rem
     }
-
-    .Player {
-        font-weight: bold;
-    }
-
 
     .input-container {
         display: flex;
@@ -280,39 +193,6 @@ channel-width=12rem;
         }
     }
 
-    .users-toggle, .channel-toggle {
-        display: none;
-    }
-
-    @media (max-width: 800px){
-        .users-toggle, .channel-toggle {
-            display: inline-block;
-
-        }
-
-        .channel-container, .users {
-            display: none;
-        }
-        .channel-container.force-show, .users.force-show {
-            display: inline-block;
-        }
-
-
-        .center-container {
-            .chat-log {
-                max-width: 100vw;
-            }
-        }
-
-        &.force-show-users {
-            .center-container {
-                .chat-log {
-                    max-width: "calc(100vw - %s)" % channel-width;
-                }
-            }
-        }
-    }
-
     .chat-line {
         overflow: hidden;
         text-overflow: ellipsis;
@@ -324,5 +204,93 @@ channel-width=12rem;
             font-family: 'Source Code Pro', monospace;
             color: #666;
         }
+    }
+}
+
+.ChatUsersList {
+    flex-basis: 12rem;
+    flex-shrink: 0;
+    flex-grow: 0;
+    overflow-y: auto;
+    user-select: none;
+
+    .user-header {
+        margin-top: 0.5rem;
+        text-transform: uppercase;
+        font-weight: bold;
+        font-size: font-size-smaller;
+        margin-bottom: 0.2rem;
+        cursor: pointer;
+        .fa-sort-numeric-asc, .fa-sort-alpha-asc  {
+            margin-left: 0.2rem;
+            width: 0.8rem;
+        }
+    }
+}
+
+.ChatChannelList {
+    .channel {
+        padding-left: 0.2rem;
+        display: flex;
+        justify-content: space-between;
+        align-items: center;
+
+        &.active {
+            themed background-color primary
+            themed color colored-background-fg
+        }
+        &.unjoined {
+            themed color shade2
+        }
+        &.unread {
+            themed color primary
+        }
+        &.mentioned {
+            themed color chat-mentions
+        }
+
+        .channel-name {
+            flex-shrink: 1;
+            flex-grow: 1;
+            cursor: pointer;
+            overflow: hidden;
+            white-space: nowrap;
+        }
+
+        .unread-count {
+            flex-shrink: 0;
+            @extend .monospace
+            font-size: 0.8rem;
+            font-style: normal !important;
+            float: right;
+            padding-left: 0.5rem;
+            padding-top: 0.10rem;
+
+            &:before {
+                content: attr(data-count);
+            }
+        }
+
+        .icon {
+            width: 15px;
+            height: 15px;
+        }
+    }
+
+    .channel:hover .unread-count:before {
+        cursor: pointer;
+        border: 1px solid transparent
+        border-radius: 0.2rem;
+        themed background-color shade4
+        themed border-color shade3
+        padding-left: 0.3rem;
+        padding-right: 0.3rem;
+        themed color fg
+    }
+
+    .channel:hover .unread-count:hover:before {
+        themed background-color shade5
+        themed border-color shade4
+        themed color fg
     }
 }
