+++ src/views/ChatView/ChatView.styl
@@ -16,10 +16,95 @@
  */
 
 
+channel-utils-height=8.4rem;
+channel-width=12rem;
+
 .ChatView {
     position: fixed;
     top: navbar-height+0.2rem;
     bottom: 0;
     left: 0;
     right: 0;
+
+    display: flex;
+    flex-direction: row;
+
+
+
+
+    .ChatChannelList, .ChatUsersList {
+        flex-basis: channel-width;
+        flex-shrink: 0;
+        flex-grow: 0;
+        width: channel-width;
+        //overflow-y: scroll;
+
+        display: flex;
+        flex-direction: column;
+        user-select: none;
+
+        flex-grow: 0;
+        flex-shrink: 0;
+        overflow-x: hidden;
+    }
+
+    .ChatChannelList {
+        border-right: 1px solid transparent;
+        themed border-right-color shade3
+    }
+
+    .ChatUsersList {
+        border-left: 1px solid transparent;
+        themed border-left-color shade3
+    }
+
+    .EmbeddedChat {
+        flex-grow: 1;
+
+        position: relative;
+        border-right: 1px solid transparent;
+        border-left: 1px solid transparent;
+        border-top: 1px solid transparent;
+        min-height: 20rem;
+        //themed border-right-color shade3
+        //themed border-color shade3
+        .input-container {
+            border-bottom: 1px solid transparent;
+        }
+    }
 }
+
+/*
+    .users-toggle, .channel-toggle {
+        display: none;
+    }
+
+    @media (max-width: 800px){
+        .users-toggle, .channel-toggle {
+            display: inline-block;
+
+        }
+
+        .channel-container, .users {
+            display: none;
+        }
+        .channel-container.force-show, .users.force-show {
+            display: inline-block;
+        }
+
+
+        .center-container {
+            .chat-log {
+                max-width: 100vw;
+            }
+        }
+
+        &.force-show-users {
+            .center-container {
+                .chat-log {
+                    max-width: "calc(100vw - %s)" % channel-width;
+                }
+            }
+        }
+    }
+    */
