+++ src/views/Tournament/Tournament.styl
@@ -82,7 +82,7 @@
         flex-wrap: wrap;
         align-items: flex-start;
 
-        .EmbeddedChat {
+        .EmbeddedChatCard {
             flex-shrink: 0;
             flex-grow: 1;
             min-width: 15rem;
